Update r.rst
Added a set of introductory sections.

Vignesh: Please see notes on the script for you to complete. These consists in:

In "About R-MGCI": Provide details on how you would prefer to receive feedback.
In "Authors" please complete the list of contributors to the tools.
In "Introduction": Please specify the packages that are needed to run the script.